backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
httpx
sentry-sdk[fastapi]
